agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.0.6",
    "strands-agents>=1.17.0",
    "bedrock-agentcore-starter-toolkit", # for CLI commands, deployment automation, development workflow support
    "valkey-glide-sync>=2.1.1",
    "mypy-boto3-bedrock-runtime>=1.42.3",
    "valkey-glide[all]>=2.2.1",
]

# Necessary for AgentCore Builder
[tool.setuptools]
py-modules = ["entrypoint", "cache_constants", "support_agent"]
